chore: move inputs to new dir
